% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{agg_y_over_x_and_avg_over_ref_id}
\alias{agg_y_over_x_and_avg_over_ref_id}
\title{Aggregate and average}
\usage{
agg_y_over_x_and_avg_over_ref_id(df, x, y, filename)
}
\arguments{
\item{df}{the tibble with indel information}
}
\value{
the invisible filename from ggsave
}
\description{
Aggregate y over x. Then take the average over all references (barcodes). The result is saved as ggplot geom_col.
}
\examples{
df <- agg_y_over_x_and_avg_over_ref_id(df)
}
