% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annote.R
\name{annote_all}
\alias{annote_all}
\title{Annote indel information}
\usage{
annote_all(df)
}
\arguments{
\item{df}{the tibble with mutation information}
}
\value{
the tibble with annoted with detailed indel information
}
\description{
Call all annotation methods on the input tibble with mutation information.
}
\examples{
df <- annote_all(df)
}
