% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annote.R
\name{annote_deletion_type}
\alias{annote_deletion_type}
\title{Annote deletion type}
\usage{
annote_deletion_type(df, mh_thres)
}
\arguments{
\item{df}{the tibble with binary indel inidicator information and micro-homology sequence}

\item{mh_thres}{the maximal mh length considered as non-micro-homology}
}
\value{
the tibble with annoted deletion type information
}
\description{
Annote deletion type (MMEJ, unilateral_top, unilateral_bottom, medial, no_deletion).
Annote extended deletion type (MMEJ_noran, unilateral_top_noran, unilateral_bottom_noran, medial_noran, MMEJ_ran, unilateral_top_ran, unilateral_bottom_ran, medial_ran, no_deletion).
}
\examples{
df <- annote_indel_type(df)
}
