% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annote.R
\name{annote_indel_size}
\alias{annote_indel_size}
\title{Annote indel size}
\usage{
annote_indel_size(df)
}
\arguments{
\item{df}{the tibble with binary indel indicator information}
}
\value{
the tibble with annoted indel size
}
\description{
Annote indel size to the input tibble (including up/down-stream indel size, deletion size, templated insertion size, random insertion size, insertion size).
}
\examples{
df <- annote_indel_size(df)
}
