% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annote.R
\name{annote_indicator}
\alias{annote_indicator}
\title{Annote binary indicator of indels}
\usage{
annote_indicator(df)
}
\arguments{
\item{df}{the tibble containing mutant information}
}
\value{
the tibble with annoted indicators
}
\description{
Annote whether up/down-stream indels actually happen.
}
\examples{
df <- annote_indicator(df)
}
