% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annote.R
\name{filter_treat}
\alias{filter_treat}
\title{Filter bad rows and select useful columns}
\usage{
filter_treat(df)
}
\arguments{
\item{df}{the tibble with annoted information}
}
\value{
the filtered tibble with selected columns
}
\description{
Filter the row with 1bp substitution. Select useful columns from annoted mutation tibble.
}
\examples{
df <- filter_treat(df)
}
