% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmej.R
\name{get_micro_homology}
\alias{get_micro_homology}
\title{Get micro homology}
\usage{
get_micro_homology(ref_end1, ref_start2, cut1, cut2, ref1, ref2)
}
\arguments{
\item{ref_end1}{the alignment end postion in the upstream reference}

\item{ref_start2}{the alignment start postion in the downstream reference}

\item{cut1}{the cleavage position in the upstream reference}

\item{cut2}{the cleavage position in the downstream reference}

\item{ref1}{the upstream reference}

\item{ref2}{the downstream reference}
}
\value{
the micro homology sequence
}
\description{
Get the micro homology around ref_end1 and ref_start2. The micro homology cannot cross the cleavage site.
}
\examples{
ref1 <- "CTTGACGCGGCGCATGCAGAAGGGCCTGGGGTGGAAGGCGTCACCGTACT"
ref2 <- "TTACCGCACTTCACGGAGATCTTGCGATGCAGGGCGGGGCTCATCTCGTGTGGCAGCTTGGCCATGCTGAAGAGCACGTAGAACATCTCGTCCACGTTGGTGTTCTTCTT"
ref_end1 <- 3
ref_start2 = 61
cut1 <- 50
cut2 <- 60
mh <- get_micro_homology(ref_end1, ref_start2, cut1, cut2, ref1, ref2)
}
