% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_control}
\alias{read_control}
\title{Read the finally control output of the naapam package}
\usage{
read_control(filename)
}
\arguments{
\item{filename}{the full path to the csv file of control}
}
\value{
a tibble with selected columns
}
\description{
The pip package naapam output a csv file of control with mutation information parsed by the bioconda package rearr. The columns are ref_id (i), cut1 (i), cut2 (i), ref_end1 (i), random_insertion (c), ref_start2 (i), count (d), stem (c), chip (c), time (i), wt (c), barcode_id (i), barcode (c), sgRNA (c), ref1 (c), ref2 (c), mh (c). This function read this file and select inportant columns.
}
\examples{
df_control <- read_control(filename)
}
