% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sgRNA_pivot}
\alias{sgRNA_pivot}
\title{Pivot with sgRNA as row}
\usage{
sgRNA_pivot(df, name, value)
}
\arguments{
\item{df}{the tibble with indel information}

\item{name}{the tidyselect column as pivot column}

\item{value}{the tidyselect column as pivot value}
}
\value{
the pivot wide table
}
\description{
Pivot with sgRNA as row, name as column and value as value. Fill missing value as 0. Repeated values are summed.
}
\examples{
df <- sgRNA_pivot(df, indel_type, count)
}
