% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sgRNA_summarise}
\alias{sgRNA_summarise}
\title{Summarise the mean along sgRNA}
\usage{
sgRNA_summarise(df, name, value)
}
\arguments{
\item{df}{the tibble with indel information}

\item{name}{the column as group}

\item{value}{the value to calculate mean}
}
\value{
the tibble with two columns: name and value
}
\description{
Summarise the mean of value along sgRNA with name as group. NA is removed. sgRNA_Pivot is applied before summarise to so that missing values are fill with 0 before taking the mean along sgRNA.
}
\examples{
df <- sgRNA_summarise(df, indel_type, count)
}
