% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annote.R
\name{annote_indel_type}
\alias{annote_indel_type}
\title{Infer indel type}
\usage{
annote_indel_type(df)
}
\arguments{
\item{df}{the tibble with binary indel inidicator information}
}
\value{
the tibble with annoted indel type information
}
\description{
Infer indel type (insertion, deletion, indel, wt) and extended indel type (wt, deletion, templated, random, temran, temdel, randel, full) from binary indel indicators.
}
\examples{
df <- annote_indel_type(df)
}
