% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_treat}
\alias{read_treat}
\title{Read the finally output of the naapam package}
\usage{
read_treat(filename)
}
\arguments{
\item{filename}{the full path to the csv file}
}
\value{
a tibble with selected columns
}
\description{
The pip package naapam output a csv file with mutation information parsed by the bioconda package rearr. This function read this file and select inportant columns.
}
\examples{
df <- read_treat(filename)
}
